test_table/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.14.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
